refactor(sandbox): move freestyle runner and docker helpers into mesh-plugin-user-sandbox
Freestyle runner (daemon-script, runner, test) migrated from apps/mesh/src/sandbox/
to packages/mesh-plugin-user-sandbox/server/runner/freestyle/ so the plugin owns
all sandbox runner implementations; apps/mesh imports it via the new
./runner/freestyle subpath export.

Docker runner reorganized into its own directory: docker.ts -> docker/runner.ts,
plus local-ingress and sweep moved alongside. Shared helpers (dev-server, handle,
inflight, lock, preview-url) extracted into a shared/ directory so freestyle and
docker runners stop duplicating them.

RunnerKind type moved from runner/index.ts to runner/types.ts (single source of
truth); plugin package.json gains ./runner/freestyle subpath export and migrates
freestyle SDKs to optionalDependencies since docker-only deploys don't need them.

No behavior change.
