+++ packages/mesh-plugin-user-sandbox/package.json
@@ -9,7 +9,14 @@
   },
   "exports": {
     "./shared": "./shared.ts",
-    "./runner": "./server/runner/index.ts"
+    "./runner": "./server/runner/index.ts",
+    "./runner/freestyle": "./server/runner/freestyle/index.ts"
+  },
+  "optionalDependencies": {
+    "@freestyle-sh/with-bun": "^0.2.12",
+    "@freestyle-sh/with-deno": "^0.0.4",
+    "@freestyle-sh/with-nodejs": "^0.2.9",
+    "freestyle-sandboxes": "^0.1.46"
   },
   "devDependencies": {
     "@types/bun": "latest",
